+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v4.0.0
+
+- Fixed (#2).
+- Added notes.
+- Added Flags.
+- Added flags support to help command.
+- Number of arguments check.
+- On log-in, password input gets hidden.
+- New Krypt Logo
+
 # v3.3.3
 
 - Fixed password strength calculation problem (#3).